backend/python-service/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
fastapi
uvicorn
python-dotenv
PyJWT==2.8.0
passlib==1.7.4

